api/requirements.txt
Flask
pytest
mock
pymongo
python-dateutil
flask-cors
numpy
pandas
requests
scipy
boto3
Pillow
opencv-contrib-python